requirements.txt
asgiref==3.7.2
attrs==23.2.0
coverage==7.4.1
Django==5.0.2
django-debug-toolbar==4.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
drf-yasg==1.21.7
flake8==5.0.4
geographiclib==2.0
geopy==2.4.1
inflection==0.5.1
install==1.3.5
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mccabe==0.7.0
packaging==23.2
pillow==10.2.0
psycopg2-binary==2.9.9
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
rpds-py==0.18.0
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
